Complete item quantity dropdown
